split out write and get fn in changelog
